% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_CurveComp_Perm.R
\name{draw_CurveComp_Perm}
\alias{draw_CurveComp_Perm}
\title{Draw DDEGs trajectories from two master.list object and also show LOESS curve fitting with permuataon}
\usage{
draw_CurveComp_Perm(
  master.list.1,
  master.list.2,
  ht.1,
  pathway,
  group.1.name,
  group.2.name,
  n.perm = 500,
  parall = F,
  pvalue.threshold = 0.05
)
}
\arguments{
\item{master.list.1, }{master.list object from group 1}

\item{master.list.2, }{master.list object from group 2}

\item{ht.1, }{TimeHeatmap object from group 1}

\item{pathway, }{GO term}

\item{group.1.name, }{characters, group 1 name}

\item{group.2.name, }{characters, group 2 name}

\item{n.perm, }{time to permutate}

\item{parall, }{run using multiple core}

\item{pvalue.threshold, }{0.05}
}
\value{
plot significant curve area
}
\description{
Draw DDEGs trajectories from two master.list object and also show LOESS curve fitting with permuataon
}
