% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_TimeHeatmap_selGO.R
\name{draw_TimeHeatmap_selGO}
\alias{draw_TimeHeatmap_selGO}
\title{Subset TimeHeatmap}
\usage{
draw_TimeHeatmap_selGO(
  time_heatmap,
  sel.go,
  master.list,
  GO.perc.thres = 0,
  nDDEG.thres = 0,
  term.width = 80,
  figure.title = "",
  save.tiff.path = NA,
  tiff.res = 100,
  tiff.width = 1500,
  tiff.height = 1500
)
}
\arguments{
\item{time_heatmap, }{TimeHeatmap object}

\item{sel.go, }{GO terms}

\item{master.list, }{master.list object}

\item{GO.perc.thres, }{0}

\item{nDDEG.thres, }{number of DDEGs, 0}

\item{term.width, }{width}

\item{figure.title, }{title}

\item{save.tiff.path, }{path}

\item{tiff.res, }{resolution}

\item{tiff.width, }{width}

\item{tiff.height, }{height}
}
\value{
draw a TimeHeatmap with selected GOs
}
\description{
Subset TimeHeatmap
}
